move groupby_agg out of DataFrame class
