+++ python/dask_cudf/dask_cudf/groupby.py
@@ -1,11 +1,154 @@
 # Copyright (c) 2020, NVIDIA CORPORATION.
+import math
+from operator import getitem
 
 import numpy as np
 import pandas as pd
 
-from dask.dataframe.core import _concat
+from dask.base import tokenize
+from dask.dataframe.core import _concat, new_dd_object
+from dask.highlevelgraph import HighLevelGraph
 
 import cudf
+
+# from dask.dataframe.groupby import DataFrameGroupBy
+# class CudfDataFrameGroupBy(DataFrameGroupBy):
+
+#     def aggregate(
+#         self,
+#         arg,
+#         split_every=None,
+#         split_out=1,
+#         dropna=True,
+#         out_to_host=False,
+#         sep="___",
+#         sort=False,
+#     ):
+#         if arg == "size":
+#             return self.size()
+
+#         if isinstance(self.obj, DataFrame) and isinstance(arg, list):
+#             _supported = {"count", "mean", "std", "var", "sum", "min", "max"}
+#             if set(arg).issubset(_supported):
+#                 index = self.index
+#                 return groupby_agg(
+#                     self,
+#                     index,
+#                     arg,
+#                     split_every=split_every,
+#                     split_out=split_out,
+#                     dropna=dropna,
+#                     out_to_host=out_to_host,
+#                     sep=sep,
+#                     sort=sort,
+#                 )
+
+#         return super().aggregate(
+#             arg, split_every=split_every, split_out=split_out)
+
+
+def groupby_agg(
+    ddf,
+    gb_cols: list,
+    agg_list: list,
+    split_every=None,
+    split_out=None,
+    dropna=True,
+    out_to_host=False,
+    sep="___",
+    sort=False,
+):
+    # Deal with default split_out and split_every params
+    if split_every is False:
+        split_every = ddf.npartitions
+    split_every = split_every or 8
+    split_out = split_out or 1
+
+    columns = [c for c in ddf.columns if c not in gb_cols]
+
+    # Only support basic aggs and mean
+    _supported = {"count", "mean", "std", "var", "sum", "min", "max"}
+    if not set(agg_list).issubset(_supported):
+        raise ValueError(
+            f"Supported aggs include {_supported} for groupby_agg API."
+        )
+
+    dsk = {}
+    token = tokenize(ddf, gb_cols, agg_list)
+    partition_agg_name = "groupby_partition_agg-" + token
+    tree_reduce_name = "groupby_tree_reduce-" + token
+    gb_agg_name = "groupby_agg-" + token
+    for p in range(ddf.npartitions):
+        # Perform groupby aggregation on each partition.
+        # Split each result into `split_out` chunks (by hashing `gb_cols`)
+        dsk[(partition_agg_name, p)] = (
+            _groupby_partition_agg,
+            (ddf._name, p),
+            gb_cols,
+            agg_list,
+            columns,
+            split_out,
+            dropna,
+            out_to_host,
+            sort,
+            sep,
+        )
+        # Pick out each chunk using `getitem`
+        for s in range(split_out):
+            dsk[(tree_reduce_name, p, s, 0)] = (
+                getitem,
+                (partition_agg_name, p),
+                s,
+            )
+
+    # Build reduction tree
+    parts = ddf.npartitions
+    widths = [parts]
+    while parts > 1:
+        parts = math.ceil(parts / split_every)
+        widths.append(parts)
+    height = len(widths)
+    for s in range(split_out):
+        for depth in range(1, height):
+            for group in range(widths[depth]):
+
+                p_max = widths[depth - 1]
+                lstart = split_every * group
+                lstop = min(lstart + split_every, p_max)
+                node_list = [
+                    (tree_reduce_name, p, s, depth - 1)
+                    for p in range(lstart, lstop)
+                ]
+
+                dsk[(tree_reduce_name, group, s, depth)] = (
+                    _tree_node_agg,
+                    node_list,
+                    gb_cols,
+                    agg_list,
+                    split_out,
+                    dropna,
+                    out_to_host,
+                    sort,
+                    sep,
+                )
+
+    # Final output partitions
+    for s in range(split_out):
+        dsk[(gb_agg_name, s)] = (
+            _finalize_gb_agg,
+            (tree_reduce_name, 0, s, height - 1),
+            gb_cols,
+            agg_list,
+            columns,
+            sep,
+        )
+
+    divisions = [None] * (split_out + 1)
+    _meta = ddf._meta.groupby(gb_cols, as_index=False).agg(agg_list)
+    graph = HighLevelGraph.from_collections(
+        gb_agg_name, dsk, dependencies=[ddf]
+    )
+    return new_dd_object(graph, gb_agg_name, _meta, divisions)
 
 
 def _make_name(*args, sep="_"):
